Hotfix 0.12.2

See merge request choerodon-c7ncd/devops-service!317

+++ src/main/java/io/choerodon/devops/domain/service/impl/HandlerC7nCertificationServiceImpl.java
@@ -28,6 +28,8 @@ import io.choerodon.devops.infra.common.util.GitUtil;
 import io.choerodon.devops.infra.common.util.TypeUtil;
 import io.choerodon.devops.infra.common.util.enums.*;
 import io.choerodon.devops.infra.dataobject.CertificationFileDO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 
 @Service
@@ -85,14 +87,16 @@ public class HandlerC7nCertificationServiceImpl implements HandlerObjectFileRela
         beforeC7nCertification
                 .forEach(certName -> {
                     CertificationE certificationE = certificationRepository.queryByEnvAndName(envId, certName);
-                    if (!CommandType.DELETE.getType().equals(certificationE.getCommandType())) {
-                        certificationE.setCommandId(certificationService
-                                .createCertCommandE(CommandType.DELETE.getType(), certificationE.getId(), userId));
-                        certificationRepository.updateCommandId(certificationE);
+                    if (certificationE != null) {
+                        if (!CommandType.DELETE.getType().equals(certificationE.getCommandType())) {
+                            certificationE.setCommandId(certificationService
+                                    .createCertCommandE(CommandType.DELETE.getType(), certificationE.getId(), userId));
+                            certificationRepository.updateCommandId(certificationE);
+                        }
+                        certificationService.certDeleteByGitOps(certificationE.getId());
+                        devopsEnvFileResourceRepository
+                                .deleteByEnvIdAndResource(envId, certificationE.getId(), ObjectType.CERTIFICATE.getType());
                     }
-                    certificationService.certDeleteByGitOps(certificationE.getId());
-                    devopsEnvFileResourceRepository
-                            .deleteByEnvIdAndResource(envId, certificationE.getId(), ObjectType.CERTIFICATE.getType());
                 });
         addC7nCertification.stream().forEach(c7nCertification -> {
             String filePath = "";
